Generalizes the iptables config test logic
Rather sloppy implementation, but gets the job done in terms of local
testing via Vagrant. These checks are for an exact match of the expected
ruleset, so any tweaks to the iptables rules in the Ansible config, even
in comments, will trigger failure here. Essentially we're only checking
implementation, rather than functionality, but that's fine for now,
since we're about to make a lot of changes to the Ansible config, and we
want to make sure we've got the system state as we want it.
